Update LAB[MB-800]_M02_Lab01_Create_company.md
The spacing was throwing somethings off and the numbers are easier to follow that the number/letter combo, added explanation of where to search for pages. It seemed a bit clumsy without an explanation of where to find the pages the instructions were directing the user to. I also added some explanation for the edit assist on some steps because it may not be intuitive to all users.